+++ spike/dev_setup.py
@@ -14,7 +14,7 @@ from codecs import encode, decode
 
 ProgramName = "SPIKE"
 VersionName = "Development version"
-VersionInfo = ["0", "99", "3"]   # Major - Minor - Micro
+VersionInfo = ["0", "99", "4"]   # Major - Minor - Micro
 
 # Major.minor.micro (int) + name eg SPIKE_2_19_5
 # N.M.L
@@ -27,6 +27,15 @@ VersionInfo = ["0", "99", "3"]   # Major - Minor - Micro
 # Release Notes in md syntax !
 release_notes="""
 # SPIKE Relase Notes
+#### 0.99.4 - April 2019
+- support for distribution via pip - you can now do `pip instal spike_py` and spike installed globally on your system.
+- still struggling with correct calibration routines for FTICR ! - proceed carefully ! -
+- phase() speeded-up by a factor 20 !
+- added a autothresh scaling to peakpicking (catching peaks "autothresh" times above the noise level - default is 3)
+- slight improvement of peak list reporting (additional key words: format="report" and format="full")
+- a bug in extract of complex 1D data-sets was corrected
+- added the figure keyword to peaklist display
+
 #### 0.99.3 - March 2019
 - Development of Interactive tools, to be used within Jupyter - *should be extended in further releases* -
     - a tool for displaying multiresolution  2D FT-ICR-MS data-sets